refactor project summary
